Read geospatial data into a Spatial RDD

Import spatial object from an external data source into a Sedona SpatialRDD.

sedona_read_shapefile: from a shapefile

sedona_read_geojson: from a geojson file

sedona_read_wkt: from a geojson file

sedona_read_wkb: from a geojson file

Usage

sedona_read_geojson(
  sc,
  location,
  allow_invalid_geometries = TRUE,
  skip_syntactically_invalid_geometries = TRUE,
  storage_level = "MEMORY_ONLY",
  repartition = 1L
)

sedona_read_wkb(
  sc,
  location,
  wkb_col_idx = 0L,
  allow_invalid_geometries = TRUE,
  skip_syntactically_invalid_geometries = TRUE,
  storage_level = "MEMORY_ONLY",
  repartition = 1L
)

sedona_read_wkt(
  sc,
  location,
  wkt_col_idx = 0L,
  allow_invalid_geometries = TRUE,
  skip_syntactically_invalid_geometries = TRUE,
  storage_level = "MEMORY_ONLY",
  repartition = 1L
)

sedona_read_shapefile(sc, location, storage_level = "MEMORY_ONLY")

Arguments

sc

A spark_connection.

location

Location of the data source.

allow_invalid_geometries

Whether to allow topology-invalid geometries to exist in the resulting RDD.

skip_syntactically_invalid_geometries

Whether to allows Sedona to automatically skip syntax-invalid geometries, rather than throwing errorings.

storage_level

Storage level of the RDD (default: MEMORY_ONLY).

repartition

The minimum number of partitions to have in the resulting RDD (default: 1).

wkb_col_idx

Zero-based index of column containing hex-encoded WKB data (default: 0).

wkt_col_idx

Zero-based index of column containing hex-encoded WKB data (default: 0).

Value

A SpatialRDD.

See also

Examples

library(sparklyr)
library(apache.sedona)

sc <- spark_connect(master = "spark://HOST:PORT")

if (!inherits(sc, "test_connection")) {
  input_location <- "/dev/null" # replace it with the path to your input file
  rdd <- sedona_read_geojson(sc, location = input_location)
}